bug fixed

+++ src/transformers/models/llama/modeling_llama.py
@@ -225,7 +225,7 @@ class LlamaAttention(nn.Module):
                 )
             attn_weights = attn_weights + attention_mask
             attn_weights = torch.max(
-                attn_weights, torch.tensor(torch.finfo(attn_weights.dtype).min, device=attn_weights.device)
+                attn_weights, torch.tensor(torch.finfo(attn_weights.dtype).min, device=attn_weights.device, dtype=attn_weights.dtype)
             )
 
         # upcast attention to fp32
